Language name fix
Although Castillian is also correct because it's a language born in Castilla (a region of Spain), the term Spanish is a better choice because this language is the language for more regions of Spain than Castilla, and also it's a language common for all Spain, even the regions that also have and use its own language. Moreover it is best know as Spanish both in Spain and outside Spain.